requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.4
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.6.0
bleach==3.1.5
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
colorama==0.4.3
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2
djangorestframework==3.12.4
entrypoints==0.3
filelock==3.0.12
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jamo==0.4.1
jedi==0.17.0
Jinja2==2.11.2
joblib==1.0.1
JPype1 @ file:///C:/Users/Msi/Downloads/JPype1-1.2.0-cp38-cp38-win_amd64.whl
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kollocate==0.0.2
koparadigm==0.10.0
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.2
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.20.1
oauthlib==3.1.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
pylint==2.5.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
redis==3.5.3
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
scipy==1.6.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
tqdm==4.57.0
traitlets==4.3.3
tweepy==3.10.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.3
virtualenv==20.4.3
wcwidth==0.2.4
webencodings==0.5.1
whitenoise==5.2.0
Whoosh==2.7.4
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
zope.interface==5.3.0
